Use border style consistently in the widgets sample
Use GetAttrs().m_defaultFlags everywhere when creating the widgets, it
was done for some but not all of them before, without any apparent
reason.

This should make setting various border styles work (for the widgets
supporting them).

+++ samples/widgets/widgets.h
@@ -115,7 +115,7 @@ struct WidgetAttributes
     wxWindowVariant m_variant;
     wxCursor m_cursor;
     // the default flags, currently only contains border flags
-    int m_defaultFlags;
+    long m_defaultFlags;
 };
 
 class WidgetsPage : public wxPanel
